requirements_common.txt
appdirs==1.4.0
Django==1.10.5
django-bootstrap3-datetimepicker==2.2.3
olefile==0.44
packaging==16.8
Pillow==4.0.0
pkg-resources==0.0.0
pyparsing==2.1.10
pytz==2016.10
six==1.10.0
